% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmax.r
\name{blockmax}
\alias{blockmax}
\title{Compute Disjoint or Sliding Block Maxima of a Univariate Time Series}
\usage{
blockmax(data, r, method = c("disjoint", "sliding"))
}
\arguments{
\item{data}{a vector of length n representing the time series}

\item{r}{the block length parameter}

\item{method}{either \code{disjoint} or \code{sliding}}
}
\value{
The time series of disjoint block maxima
(a vector of length \eqn{\lfloor n/r \rfloor})
or sliding block maxima
(a vector of length \eqn{n-r+1})
}
\description{
For a given univariate time series of length n and block size
parameter r (\eqn{1 \le r \le n}), \code{blockmax} computes either the time
series of disjoint (non-overlapping) or sliding (overlapping) block maxima.
}
\examples{
x <- runif(200)
blockmax(x, 20, "disjoint")
blockmax(x, 20, "sliding")
}
