% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_concat_years.R
\name{bm_concat_years}
\alias{bm_concat_years}
\title{Compute sliding block maxima of the time series obtained by concatenating
observations from two arbitrary years.}
\usage{
bm_concat_years(dailydata, ind1, ind2, blcksz = 90)
}
\arguments{
\item{dailydata}{dataframe with hourly/daily ... observations in column named 'obs',
and the Index for the observation (e.g. the year/season in which the observation was made)
in column named 'Index'}

\item{ind1}{Observations with that index make up the first part of the
concatenated series}

\item{ind2}{Observations with that index make up the first part of the
concatenated series}

\item{blcksz}{The blocksize for computing sliding block maxima.}
}
\value{
a numeric vector containing the sliding block maxima
}
\description{
Compute sliding block maxima of the time series obtained by concatenating
observations from two arbitrary years.
}
\examples{
df <- data.frame(obs = evd::rgpd(30*5), Index = rep(1:5, each = 30))
bm_concat_years(df, 2, 4, blcksz = 30)
}
