% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification_functions.R
\name{check_fun}
\alias{check_fun}
\title{Check function for evaluating an estimated quantile}
\usage{
check_fun(u, p)
}
\arguments{
\item{u}{Difference of observed maximum in a test set and estimated p*100\% quantile}

\item{p}{Probability with which the difference is weighted}
}
\value{
value of check function
}
\description{
Check function for evaluating an estimated quantile
}
\examples{
check_fun(2, 0.9)
}
