% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_boot_sl.R
\name{ci_student_boot_sl}
\alias{ci_student_boot_sl}
\title{Estimate RLs and CIs based on sliding blocks}
\usage{
ci_student_boot_sl(
  yy,
  blcksz,
  Kblock = c(4, 8, 10),
  B = 200,
  temp.cov = NULL,
  type = "shift",
  scale.link = make.link("identity"),
  Tyrl = c(50, 100),
  ref_gmst = c(0.8, 0.9),
  varmeth = "V2",
  looplastblock = TRUE,
  reltol = 1e-09
)
}
\arguments{
\item{yy}{Numeric vector of data with daily resolution}

\item{blcksz}{The blocksize parameter}

\item{Kblock}{Value(s) of number of blocks making up one K-block within the block-bootstrap}

\item{B}{Number of bootstrap iterations}

\item{temp.cov}{temporal covariate, can be NULL}

\item{type}{Either one of 'stationary' (no trend), 'shift' (linear trend in location parameter)
or 'scale' (trend in location and scale parameter such that their ratio is constant over time)}

\item{scale.link}{link function for the scale parameter}

\item{Tyrl}{The return periods for which RLs are estimated}

\item{ref_gmst}{Vector of reference values of the temporal covariate for which
RLs are estimated. Ignored when type is 'stationary'.}

\item{varmeth}{The method used to estimate the covariance matrix of GEV parameters: 'V' or 'V2'.}

\item{looplastblock}{logical; wether last block is looped}

\item{reltol}{Passed to optim}
}
\value{
A tibble containing for each combination of \code{Year}, \code{ref_gmst} and
\code{Kblocks} the following values:
\describe{
\item{rl}{The RL that was estimated on the concatenated Kblock-sample.
Only differs from the RL that was estimated on the full sliding BM sample when
\code{looplastblock = TRUE}.}
\item{var_hat}{The estimated Variance of the RL estimation.}
\item{q0975, q0025}{The 97.5\%- and 2.5\%- quantiles of the bootstrapped studentized RLs.}
\item{lower, upper}{The lower and upper bound of the confidence interval.}
\item{further parameters}{Denoted by  \code{mu0, sigma0, gamma, alpha} for 'scale',
\code{loc0, scale0, shape, TempLoc1} for 'shift' and by
\code{loc, scale, shape} for 'stationary', giving the estimated GEV parameters as estimated
on the Kblock-sample.}
}
}
\description{
Estimate RLs and respective confidence intervals of GEV fits based on
sliding Block maxima, with possibility to consider a shift or a scaling of maxima
with respect to a temporal covariate.
Confidence intervals are obtained via block bootstrap.
}
\examples{
blcksz <- 90
ny <- 50
## with a shift in location
yy <- evd::rgpd(ny*blcksz, shape = 0.2) + 2*rep(1:ny/ny, each = blcksz)
ci_student_boot_sl(yy, blcksz = 90, temp.cov = rep(1:ny/ny, each = blcksz),
B = 100, type = "shift", ref_gmst = c(0.5, 0.95))

## observations are scaling w.r.t. GMST
yy <- evd::rgpd(ny*blcksz, shape = 0.2)*exp(2/7.3*rep(1:ny/ny, each = blcksz))
ci_student_boot_sl(yy, blcksz = 90, temp.cov = rep(1:ny/ny, each = blcksz),
B = 100, type = "scale", ref_gmst = c(0.5, 0.95))

}
