% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conc_bm.R
\name{compute_conc_bm}
\alias{compute_conc_bm}
\title{Compute concatenated block maxima}
\usage{
compute_conc_bm(dailydata, nlo = 3, blcksz)
}
\arguments{
\item{dailydata}{dataframe with hourly/daily ... observations in column named 'obs',
and an index for each observation (e.g. the year/season in which the observation was
made) in column named 'Index'}

\item{nlo}{number of years that make up the test set}

\item{blcksz}{blocksize for computing sliding block maxima}
}
\value{
tibble with column 'ind1' giving the first index of observations making up
the concatenated time series; column 'ind2' giving the second index of observations
making up the concatenated time series; column 'newslbm': a list
of vectors containing the sliding block maxima of the concatenated time series

Note that 'ind1' is allowed to be larger than 'ind2'
}
\description{
Compute sliding block maxima for all possible compositions of years
arising from cross validation with test set of size given in 'nlo'
}
\examples{
df <- data.frame(obs = evd::rgpd(30*50), Index = rep(1:50, each = 30))
compute_conc_bm(df, nlo = 3, blcksz = 30)
}
