% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification_functions.R
\name{compute_qs}
\alias{compute_qs}
\title{Compute the quantile score for an estimated quantile}
\usage{
compute_qs(obs, quant.est, p)
}
\arguments{
\item{obs}{vector of observations from a test set}

\item{quant.est}{value of estimated p*100\% quantile}

\item{p}{probability}
}
\value{
a numeric value giving the quantile score
}
\description{
Compute the quantile score for an estimated quantile
}
\examples{
set.seed(1)
          train <- evd::rgev(100)
          estpars <- evd::fgev(train[1:97])$estimate
          quanthat <- evd::qgev(0.99, loc = estpars[1], scale = estpars[2],
           shape = estpars[3])
           compute_qs(train[98:100], quanthat, 0.99 )

}
