% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification_functions.R
\name{compute_qsi}
\alias{compute_qsi}
\title{Compute the Quantile Skill Index (QSI) of new model vs reference model}
\usage{
compute_qsi(qs.new.model, qs.ref.model)
}
\arguments{
\item{qs.new.model}{quantile score of the new model}

\item{qs.ref.model}{quantile score of the reference model}
}
\value{
numeric value between -1 and 1. Values smaller than 0 indicate that
the reference model performs better, while values larger than 0 indicate that
the new model performs better
}
\description{
Compute the Quantile Skill Index (QSI) of new model vs reference model
}
\examples{
set.seed(1)
train <- evd::rgev(500)
test <- evd::rgev(30)
estpars1 <- evd::fgev(train)$estimate
quanthat1 <- evd::qgev(0.99, loc = estpars1[1], scale = estpars1[2],
                     shape = estpars1[3])
estpars2 <- evd::fgev(train, loc = 0 )$estimate # fix location to true value
quanthat2 <- evd::qgev(0.99, loc = 0, scale = estpars2[1],
                     shape = estpars2[2])
qs1 <- compute_qs(test, quanthat1, 0.99 )
qs2 <- compute_qs(test, quanthat2, 0.99 )
compute_qsi(qs2, qs1)
}
