% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_boot_sl.R
\name{compute_rl}
\alias{compute_rl}
\title{Compute RL of fitted GEV distribution}
\usage{
compute_rl(theta, Tyrl, type, ref_gmst = NULL)
}
\arguments{
\item{theta}{Estimated GEV parameters}

\item{Tyrl}{The return periods for which the RLs are computed}

\item{type}{The model: 'stationary'. 'shift' or 'scale'}

\item{ref_gmst}{The reference value of the temporal covariate for which the
RL is computed. Ignored when \code{type = "stationary"}.}
}
\value{
A tibble containg the estimated RL for each combination of \code{Tyrl} and
\code{ref_gmst}.
}
\description{
Compute RL of fitted GEV distribution
}
\examples{
compute_rl(theta = c(loc0 = 7, scale0 = 2, shape = 0.2, TempLoc1 = 2),
Tyrl = c(50, 100), type = "shift", ref_gmst = c(0.5, 0.95))

}
