% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_fit_eff.R
\name{cv_fit_eff}
\alias{cv_fit_eff}
\title{Compute parameter estimates and quantile score on a test set}
\usage{
cv_fit_eff(
  djbm,
  slbm,
  conc_bm,
  leave_out,
  quants = 1 - 1/c(2, 5, 10, 20, 50, 100, 200),
  method = "both",
  fixpar = NULL
)
}
\arguments{
\item{djbm}{dataframe with columns 'obs' and 'Index' containing observed disjoint block
maxima and the year they were observed in, respectively. Needs to be passed
also when only using sliding blocks method to obtain the test set for cross-validation.}

\item{slbm}{dataframe  with columns 'sldata' and 'Index' containing observed sliding block
maxima and an Index for the block maxima, e.g. the year/season the first observation
of the respective sliding block was observerd in. With this Index, the
observations that are left out during cross-validation are specified. For example, if you want
to leave out all observations that were made in Years 1999, 2000, 2001, when computing
sliding block maxima, the column 'Index' must contain the year in which the first observation of the
sliding block was observed. When passing the years 1999, 2000, 2001 to 'leave_out', then
all observations from these years are left out, and sliding block maxima for the now consecutive
years (1998 is followd by 2002) are assembled from 'conc_bm'}

\item{conc_bm}{tibble containing sliding block maxima obtained from concatenating
observations from years/seasons... that are usually not consecutive, as computed from the function
'compute_conc_bm'}

\item{leave_out}{The indices you want to leave out from the training data; observations
with these indices make up the test data}

\item{quants}{The quantiles which you want to estimate and compute the quantile score for}

\item{method}{Specifies which estimation method to use: must be
one of "disjoint", "sliding" or "both" (default).}

\item{fixpar}{Optional: when applying the sliding block method, you can
fix one of the parameters to the respective estimated value obtained from the disjoint block method.
For the default (NULL) no parameter is fixed. Other options are "loc", "scale" or c("loc", "scale")
for fixing location, scale or both location and scale parameter. This overrides the choice of 'method'
and returns values of the disjoint method and the sliding method with fixed values if not NULL.}
}
\value{
tibble containing the estimated parameters, the estimation method, the
estimated quantiles, the test set as well as the corresponding quantile scores.
}
\description{
Estimate quantiles based on sliding and/or disjoint blocks and
efficiently evaluate estimates based on the quantile score of a subset of the data
}
\examples{
df <- data.frame(obs = evd::rgpd(30*50), Index = rep(1:50, each = 30))
cv_fit_eff(data.frame( obs = blockmax(df$obs, r = 30, "disjoint"),  Index = 1:50),
data.frame( sldata = blockmax(df$obs, r = 30, "sliding"), Index = rep(1:50, each = 30)[1:(49*30 +1)]),
conc_bm = compute_conc_bm(df, nlo = 3, blcksz = 30), leave_out = c(4, 35, 42))

}
