% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyPrec.R
\docType{data}
\name{dailyPrec}
\alias{dailyPrec}
\title{Daily cumulated precipitation amounts from Aachen}
\format{
A data frame with 38440 rows and 2 variables:
\describe{
\item{cumPrec}{daily precipitation height in mm}
\item{Date}{Date of observation}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/}
}
\usage{
dailyPrec
}
\description{
A dataset containing the cumulated precipitation amounts as observed in Aachen,
along with the day they were observed.
}
\keyword{datasets}
