% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailyTempNRW.R
\docType{data}
\name{DailyTempNRW}
\alias{DailyTempNRW}
\title{Daily Temperature Series at 9 stations in NRW.}
\format{
A list with components
\describe{
\item{TimeSeries}{ tibble (8,188 x 10): the first column \code{Datetime} contains
the date of the observation, the columns 2 to 9 the observed daily maximum temperature
at the respective station (see column names) as made during summer months (June, July, August)
}
\item{SpatialCov}{Spatial covariates such as Station ID, Station name, elevation,
latitude, longitude.}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/}
}
\usage{
DailyTempNRW
}
\description{
Daily Temperature Series at 9 stations in NRW.
}
\details{
The data was downloaded from the DWD.
}
\keyword{datasets}
