% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fun_wtrend.R
\name{est_var_chain}
\alias{est_var_chain}
\title{Estimate covariance matrix based on chain rule}
\usage{
est_var_chain(
  orig_slbm,
  est_par,
  blcksz,
  temp.cov = NULL,
  type = "shift",
  varmeth = "both",
  ...
)
}
\arguments{
\item{orig_slbm}{Sample of sliding block maxima}

\item{est_par}{estimated parameter vector}

\item{blcksz}{blocksize parameter}

\item{temp.cov}{vector of temporal covariate, should have same length as sample}

\item{type}{either \code{shift} or \code{scale}.
WARNING: for the \code{shift}-model I'm not sure anymore whether the scaling
part is missing or if it's hidden somewhere in the computations. Should be checked.}

\item{varmeth}{the method for estimating the covariance of overlapping observations,
either \code{V} or \code{V2}}

\item{...}{Further parameters. When \code{type = "scale"},
\code{rel_trend = TRUE} or \code{rel_trend = FALSE} needs to be specified here.}
}
\value{
A list containing the covariance matrices estimated with the chosen methods
}
\description{
Estimate covariance matrix based on chain rule
}
\examples{
### simulate some data with a linear trend
xx <- evd::rgpd(90*100, shape = 0.2) + 2*rep(1:100/100, each = 90)
## temporal covariate for sliding BM
temp_cvrt <- rep(1:100/100, each = 90)[1:(90*100 - 90 +1)]

## compute sample of uniuqe sliding BM
bms <- get_uniq_bm(xx, 90, temp_cvrt = temp_cvrt, looplastblock = FALSE)

## full sample of sliding for estimaing the covariance matrix
slbm <- blockmax(xx, r = 90, "sliding")
estim <- fit_gev_univ(data = bms, type = "shift", hessian = TRUE)

est_var_chain(slbm, est_par = estim, blcksz = 90, temp.cov = temp_cvrt,
type = "shift", varmeth = "V2")

}
