% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fun_wtrend.R
\name{est_var_univ}
\alias{est_var_univ}
\title{GEV parameter covariance matrix for sliding blocks ML estimator}
\usage{
est_var_univ(
  orig_slbm,
  est_par,
  blcksz,
  temp.cov = NULL,
  type = "shift",
  varmeth = "both",
  chain = TRUE,
  ...
)
}
\arguments{
\item{orig_slbm}{Full sample of sliding block maxima}

\item{est_par}{Output of \code{\link{fit_gev_univ}}.}

\item{blcksz}{The blocksize parameter.}

\item{temp.cov}{The temporal covariate (must be same length as \code{orig_slbm}).}

\item{type}{Either 'shift', 'scale' or 'stationary', depending if one assumes
that obervations shift, scale or do not change w.r.t. the covariate or stationary
(for more details, see
\code{\link[slbm]{fit_gev_univ}}).}

\item{varmeth}{One of 'V', 'V2' or 'both'. Determines the method used for estimating
the covariance matrix.}

\item{chain}{logical; whether the estimator for the covariance matrix should be one obtained
from the chain rule. Only relvant for the non-stationary models.}

\item{...}{A further argument that needs to be passed when \code{type = scale}, specifying the
parametrisation of the model.}
}
\value{
A list with components \code{V} and/or \code{V2} containing estimated covariance matrices.
}
\description{
GEV parameter covariance matrix for sliding blocks ML estimator
}
\examples{
### simulate some data with a linear trend
}
