% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fun_wtrend.R
\name{est_var_wtrend}
\alias{est_var_wtrend}
\title{GEV parameter covariance matrix for sliding blocks}
\usage{
est_var_wtrend(
  orig_slbm,
  est_par,
  blcksz,
  temp.cov,
  type = "shift",
  varmeth = "both"
)
}
\arguments{
\item{orig_slbm}{Full sample of sliding block maxima}

\item{est_par}{Output of \code{\link{fit_gev_univ_trend()}}.}

\item{blcksz}{The blocksize used.}

\item{temp.cov}{The temporal covariate (must be same length as \code{orig_slbm}).}

\item{type}{Either 'shift' or 'scale', depending if one assumes that obervations are
shifted or scaled w.r.t. the covariate (for more details, see
\code{\link{fit_gev_univ_trend()}}).}

\item{varmeth}{One of 'V', 'V2' or 'both'. Determines the method used for estimating
the covariance matrix.}
}
\value{

}
\description{
GEV parameter covariance matrix for sliding blocks
}
\examples{
### simulate some data with a linear trend
xx <- evd::rgpd(90*100, shape = 0.2) + 2*rep(1:100/100, each = 90)
## temporal covariate for sliding BM
temp_cvrt <- rep(1:100/100, each = 90)[1:8911]

## compute sample of uniuqe sliding BM
yy <- data.frame( Obs = xx, Station = "X1")
bms <- get_uniq_bm(yy, 90, temp_cvrt = temp_cvrt, looplastblock = FALSE)
bms <- bms[1, ]$uniq_data[[1]]

## full sample of sliding for estimaing the covariance matrix
slbm <- blockmax(xx, r = 90, "sliding")
estim <- fit_gev_univ_trend(data = bms, type = "shift", hessian = TRUE)
est_var_wtrend(slbm, est_par = estim, blcksz = 90, temp.cov = temp_cvrt,
varmeth = "both", type = "shift")
}
