% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homfit_functions.R
\name{fgev_hom}
\alias{fgev_hom}
\title{Maximum-Likelihood fitting of the GEV distribution under homogeneity constraint}
\usage{
fgev_hom(
  dat,
  start_vals = NULL,
  scale.link = make.link("log"),
  method = "BFGS",
  ...
)
}
\arguments{
\item{dat}{A matrix or data frame. Each column contains values for one location.
May also contain missing values.}

\item{start_vals}{Optional: a vector containing initial values for the log Likelihood
which is to be optimised. If omitted, initial values are estimated.
See below for further details.}

\item{scale.link}{The link function that is used for modelling the scale parameter.
Only used during optimisation.}

\item{method}{Optimisation method that is passed to \code{\link[stats]{optim}}.}

\item{...}{Further arguments that are passed to the \code{control} argument of \code{\link[stats]{optim}}.}
}
\value{
Returns a list with the following components:
\describe{
\item{mle}{A 3 x d dimensional matrix containing the estimated location,
scale and shape parameter at each station.}
\item{mledisp}{The estimated dispersion parameter, i.e. the ratio of the estimated
location and scale parameters.}
\item{conv}{The convergence code as obtained from \code{optim}.}
\item{counts}{A two-element integer vector giving the number of calls to
the negative log-Likelihood and its gradient, respectively, as obtained from
\code{optim}.}
\item{nllh}{Value of the minimised negative log-Likelihood.}
}
}
\description{
Maximum-Likelihood fitting of the GEV distribution under homogeneity
constraint, also known as the 'Indexflood'-assumption.
For (approximately) GEV distributed observations from several locations in space
it is assumed that the shape parameter as well as the ratio
of location and scale parameters (the \strong{dispersion} parameter) is constant
in space. To be precise, let
\eqn{X_{i,j}, i = 1, \ldots, n; j = 1, \ldots,d } denote the observations
from \eqn{d} stations. It is assumed that
\eqn{ X_{i,j} \sim GEV(\mu(j), \sigma(j), \gamma)},
where \eqn{ \mu(j) / \sigma(j) = \delta }, which does not depend on \eqn{d}.
}
\details{
If no starting values are provided, the initial values are computed as
follows: First, GEV parameters are estimated stationwise.
The initial dispersion parameter
is set to the mean of stationwise dispersion parameters, the initial shape
parameter is set to the mean of stationwise shape parameters and the initial scale
parameters are the stationwise estimates.

The scale parameters that are returned are the actual scale parameters.
E.g., when using \eqn{link(\sigma) = \tilde{\sigma}} as link function,
\eqn{\sigma =  linkinv(\tilde{\sigma})} is returned.
}
\examples{
xx <- sapply(1:5, function(x) evd::rgev(30, loc = 2*x, scale = x, shape = 0.2))
fgev_hom(xx)

}
