% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gev_univ_dj.R
\name{fit_gev_univ_dj}
\alias{fit_gev_univ_dj}
\title{GEV fit to disjoint block maxima}
\usage{
fit_gev_univ_dj(
  data,
  temp_cvrt,
  type,
  method = "BFGS",
  maxiter = 200,
  hessian = FALSE,
  rel_trend = FALSE,
  return_cov = FALSE
)
}
\arguments{
\item{data}{vector of block maxima}

\item{temp_cvrt}{vector of temporal covariate}

\item{type}{one of 'stationary', 'shift', 'scale'}

\item{method}{passed to optim}

\item{maxiter}{passed to optim}

\item{hessian}{logical; whether to return hessian}

\item{rel_trend}{refers to parametrisation of scale model, therefore only
relevant when 'type == "shift"'}

\item{return_cov}{logical; whether to return a covariance matrix}
}
\value{
A list with estimated parameter values, convergence code etc., and
covariance matrix (if desired)
}
\description{
GEV fit to disjoint block maxima
}
\examples{
# bla
}
