% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aggregate.R
\name{fun_aggregate}
\alias{fun_aggregate}
\title{Compute Intensity duration process and yearly block maxima for one duration $d$}
\usage{
fun_aggregate(data, d, resolution = "hourly", seasonlength = 365)
}
\arguments{
\item{data}{a data frame or tibble of precipitation data, with a column named
\itemize{
\item datetime: containing the date-time of the observation as POSIXct, POSIXlt object
\item prec: containing the observed precipitation amount
}}

\item{d}{the duration for which you want to compute the intensity process.
It is the number $d$ of observations over which the the moving average
of precipitation is computed, computed within distinct years.}

\item{resolution}{The resolution of the data, must be one of "hourly" or "daily"}

\item{seasonlength}{The length of the season of the input data. E.g., if observations
are from the whole year, it is 365; if observations are from JJA only, it is 92.}
}
\value{
returns a tibble with colum
\itemize{
\item aggs: tibble of aggregated observations, Year they belong to and the duration d
\item djbm: tibble of annual maxima of the intensity duration process, along with
corresponding year and duration
}
}
\description{
Compute Intensity duration process and yearly block maxima for one duration $d$
}
\examples{
dates <- seq(as.POSIXct("2000-01-01 00:00:00"),
as.POSIXct("2001-12-31 23:00:00"),by = 'hour')
prec <- rgamma(length(dates), shape =0.1)
ExampleData <- data.frame(datetime = dates, prec = prec)

fun_aggregate(ExampleData, 120)

}
