% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agg_bm.R
\name{get_agg_bm}
\alias{get_agg_bm}
\title{Compute disjoint and sliding block maxima sample of ID process}
\usage{
get_agg_bm(data, ds, resolution = "hourly")
}
\arguments{
\item{data}{a data frame or tibble of precipitation data, with a column named
\itemize{
\item datetime: containing the date-time of the observation as POSIXct, POSIXlt object
\item prec: containing the observed precipitation amount
}}

\item{ds}{the durations for which you want to compute the ID process and its annual
maxima}

\item{resolution}{the resolution of the observed data}
}
\value{
a tibble with columns duration, blcksz (the blocksize for the sliding block maxima sample,
depending on the duration), djbm (list of tibbles of disjont block maxima),
slbm ( list of tibbles og sliding block maxima) (one tibble for each duration)
}
\description{
Compute disjoint and sliding block maxima sample of ID process
}
\examples{
dates <- seq(as.POSIXct("2000-01-01 00:00:00"),
as.POSIXct("2001-12-31 23:00:00"),by = 'hour')
prec <- rgamma(length(dates), shape = 0.1)
ExampleData <- data.frame(datetime = dates, prec = prec)

get_agg_bm(ExampleData, ds = c(1,2,4,8,16, 24, 48))
}
