% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blcksz.R
\name{get_blcksz}
\alias{get_blcksz}
\title{Compute blocksize for duration d}
\usage{
get_blcksz(d, resolution = "hourly")
}
\arguments{
\item{d}{the duration}

\item{resolution}{resolution of measurements, must be one of "hourly" or "daily"}
}
\value{
returns the number of observations from an intensity duration process
that fall into one year, when one has hourly observations on 365 days
}
\description{
Compute blocksize for duration d
}
\examples{
get.blcksz(4)
}
