% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bm_data_qs.R
\name{get_bm_data_qs}
\alias{get_bm_data_qs}
\title{Get dataframes for use in quantile score functions}
\usage{
get_bm_data_qs(data, blcksz, return_sldata = TRUE, return_djdata = TRUE)
}
\arguments{
\item{data}{Dataframe or tibble with variables named 'obs' and 'Index' containing
the observations and the year/season the observation was made in.}

\item{blcksz}{The blocksize to compute blockmaxima}

\item{return_sldata}{logical; whether to return dataframe containing sliding data}

\item{return_djdata}{logical; whether to return dataframe containing disjoint data}
}
\value{
list with elements djbm and slbm
}
\description{
Get dataframes for use in quantile score functions
}
\examples{

bm_data_qs(data.frame( obs = rnorm(50*30), Index = rep(1:50, each = 30)),
blcksz = 30)
}
