% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unique_slbm.R
\name{get_uniq_bm}
\alias{get_uniq_bm}
\title{Compute stationwise unique sliding block maxima and their frequency}
\usage{
get_uniq_bm(data, blcksz, temp_cvrt = NULL, looplastblock = FALSE)
}
\arguments{
\item{data}{A tibble with columns \code{Station} and \code{Obs}, where \code{Obs} contains
measurements of the variable of interest at the Station referenced by \code{Station}.
If dealing with observations from one station only (univariate case), it can also be a
vector containing the observations.}

\item{blcksz}{The blocksize for which block maxima are computed.}

\item{temp_cvrt}{Optional; a numeric vector of length \code{nrow(data) - blcksz +1}
containing values of a temporal
covariate. The specific length is necessary in order to match each
sliding block maximum to a value of the temproal
covariate. If too long, the vector will be cut off accordingly.
To be provided if you assume the distribution of your
data to be non-stationary and want to fit parameters accordingly.
@param looplastblock logical; whether or not the first disjoint block of observations is
concatenated to the last disjoint block of observations when computing the sample of sliding
block maxima. Because of stationarity issues, this should only be used when
sample is assumed to be stationary.}
}
\value{
Returns a nested tibble with columns \code{Station} and \code{uniq_data}:
for each station, \code{uniq_data} contains a tibble with columns \code{slbm}, \code{temp_cvrt}
(if considered)
and \code{n}, with the unqiue values of sliding block maxima (and temporal covariate)
along with the respective frequency.
In the univariate case, a tibble containing the unique values and their frequency
(and, if considered, the corresponding value of the temporal covariate) is returned.
}
\description{
Compute stationwise unique sliding block maxima and their respective
frequency from observations from several locations in space over time.
If the distribution is assumed to be non-stationary, the frequencies where
block maxima and temporal covariate coincide are computed.
}
\examples{
##### generate some data #####
blcksz <- 90
xx <- sapply(c(0.1, 0.4), function(x) evd::rgpd(10*blcksz, shape = x))

yy <- data.frame(xx) \%>\% tidyr::pivot_longer( 1:2, names_to = "Station", values_to = "Obs")
##############################

# define a temporal covariate that is constant over a block of length blcksz
temp_cvrt <- rep(1:10/10, each = blcksz)[1:(9*blcksz + 1)]

##### apply function #####
bms <- get_uniq_bm(yy, blcksz, temp_cvrt = temp_cvrt )
}
