% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllh_spat_temp_sl.R
\name{nll_spat_temp_sl_hom}
\alias{nll_spat_temp_sl_hom}
\title{Negative (pseudo) log-Likelihood for spatial GEV model with covariates under
homogeneity constraint.}
\usage{
nll_spat_temp_sl_hom(
  params,
  scale.sp.form,
  scale.temp.form = NULL,
  dataprep,
  spat.cov,
  scale.link = make.link("log")
)
}
\arguments{
\item{params}{Named vector of parameters for which the negative log-Likelihood
is to be computed. See details.}

\item{scale.sp.form}{R formula definining the spatial model for the scale parameter.}

\item{scale.temp.form}{R formula definining the temporal trend for the scale parameter.}

\item{dataprep}{Prepared data of sliding block maxima and design matrices
as obtained from \link{prep4spatml}.}

\item{spat.cov}{A data frame containing the spatial covariates used in the formulations
of the spatial formulas.}

\item{scale.link}{The link function that is used for modelling the scale parameter.}
}
\value{
The value of the negative log-Likelihood (a single numeric value).
}
\description{
Computes the negative (pseudo) log-Likelihood for the spatial GEV model with
parameters dependending on spatial (and/or temporal) covariates under
the assumption of constant dispersion (location/scale) parameter.
}
\examples{
data("ExampleData")
yy <- data.frame(ExampleData) \%>\%
      tidyr::pivot_longer( 1:8, names_to = "Station", values_to = "Obs")
set.seed(3)
spatial_cvrt <- data.frame(lat  = seq(0, 8, length = 8),
   lon = runif(8), ele = runif(8))
##############################
bms <- get_uniq_bm(yy, 90)
prepdata <- prep4spatml(loc.sp.form = ~ lon + lat + ele,
              scale.sp.form = ~ lon +lat + ele, data = bms, spat.cov = spatial_cvrt)

params  <- c("disp" = 2, "scale0" = 120,
               "scale1" = -50, "scale2" = 5, "scale3" = -60, "shape" = 0.4)

nll_spat_temp_sl_hom(params = params,
          scale.sp.form = ~ lon + lat + ele,
          dataprep = prepdata, spat.cov = spatial_cvrt)

}
