% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllh_spat_temp_sl.R
\name{nll_spat_temp_sl_prep}
\alias{nll_spat_temp_sl_prep}
\title{Negative (pseudo) log-Likelihood for spatial GEV model with covariates}
\usage{
nll_spat_temp_sl_prep(
  params,
  loc.sp.form,
  scale.sp.form,
  loc.temp.form = NULL,
  scale.temp.form = NULL,
  dataprep,
  spat.cov,
  scale.link = make.link("log")
)
}
\arguments{
\item{params}{Named vector of parameters for which the negative log-Likelihood
is to be computed. See details.}

\item{loc.sp.form}{R formula definining the spatial model for the location parameter.}

\item{scale.sp.form}{R formula definining the spatial model for the scale parameter.}

\item{loc.temp.form}{R formula definining the temporal trend for the location parameter.}

\item{scale.temp.form}{R formula definining the temporal trend for the scale parameter.}

\item{dataprep}{Prepared data of sliding block maxima and design matrices
as obtained from \link{prep4spatml}.}

\item{spat.cov}{A data frame containing the spatial covariates used in the formulations
of the spatial formulas.}

\item{scale.link}{The link function that is used for modelling the scale parameter.}
}
\value{
The value of the negative log-Likelihood (a single numeric value).
}
\description{
Computes the negative (pseudo) log-Likelihood for the spatial GEV model with
parameters
dependending on spatial (and/or temporal) covariates.
}
\examples{

data("ExampleData")
data("GMST")
tempcv <- GMST \%>\% dplyr::filter(Year \%in\% c(1980:2019))
tempcv <- rep(tempcv$smoothedGMST, each = 90)[1:(39*90 +1)]
tempcv <- data.frame(GMST = tempcv)
set.seed(3)
spatial_cvrt <- data.frame(lat  = seq(0, 8, length = 8),
   lon = runif(8), ele = runif(8))
yy <- data.frame(ExampleData) \%>\%
      tidyr::pivot_longer( 1:8, names_to = "Station", values_to = "Obs")
##############################
bms <- get_uniq_bm(yy, 90, temp_cvrt = tempcv$GMST)
prepdata <- prep4spatml(loc.sp.form = ~ lon + lat + ele,
              scale.sp.form = ~ lon +lat,
              loc.temp.form = ~ GMST, data = bms, spat.cov = spatial_cvrt)

params  <- c("loc0" = 18, "loc1" = -1, "loc2" = 0.4, "loc3" = 0.02, "scale0" =0.2,
               "scale1" = 0.1, "scale2" = 0.1, "shape" = 0.3, "tempLoc1" = 1)

nll_spat_temp_sl_prep(params = params, loc.sp.form = ~ lon + lat + ele,
          scale.sp.form = ~ lon + lat, loc.temp.form = ~ GMST,
          dataprep = prepdata, spat.cov = spatial_cvrt,
          scale.link = make.link("log"))
}
