% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllh_sl_d.R
\name{nllh.sl.d}
\alias{nllh.sl.d}
\title{Negative log-Likelihood for sliding BM of ID process}
\usage{
nllh.sl.d(
  parvec,
  bm_uni,
  sigma0link = make.link("log"),
  mult_sc = TRUE,
  dur_offset = FALSE,
  int_offset = FALSE
)
}
\arguments{
\item{parvec}{Vector of parameter values}

\item{bm_uni}{Dataframe containing the unqiue values of sliding block maxima along with
their frequency of appearance, for possibly several values of the duration $d$, as
specified in column 'duration'.}

\item{sigma0link}{The link function for \eqn{\sigma_0}. Default is \code{make.link("log")}.}

\item{mult_sc}{Whether or not you pass a multi-scaling parameter.}

\item{dur_offset}{Whether or not you pass a duration offset parameter.}

\item{int_offset}{Whether or not you pass an intensity offset parameter.}
}
\value{
The value of the negative log-Likelihood at the given parameters.
}
\description{
Negative log-Likelihood for sliding BM of ID process
}
