% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unique_slbm.R
\name{prep4spatml}
\alias{prep4spatml}
\title{Prepare sliding block maxima data for ML fitting}
\usage{
prep4spatml(
  loc.sp.form,
  scale.sp.form,
  loc.temp.form = NULL,
  scale.temp.form = NULL,
  data,
  spat.cov
)
}
\arguments{
\item{loc.sp.form}{R formula definining the spatial model for the location parameter.}

\item{scale.sp.form}{R formula definining the spatial model for the scale parameter.}

\item{loc.temp.form}{R formula definining the temporal trend for the location parameter.}

\item{scale.temp.form}{R formula definining the temporal trend for the scale parameter.}

\item{data}{A nested tibble as obtained from \link{get_uniq_bm}.}

\item{spat.cov}{A data frame containing the spatial covariates used in the formulations
of the spatial formulas.}
}
\value{
A list containing the following components:
\describe{
\item{data}{The original data that was put in. If there is a trend in either
location and/or scale parameter (i.e. when the corrsesponding formulas are not NULL),
data contains an additional column \code{MatLocTemp} or \code{MatScaleTemp} (or both)
containing the design matrices for the temporal
trends as generated by \code{model.matrix()} (these are vectors of the same length
as the matrices in \code{uniq_data}).}
\item{MatLocSp}{The design matrix for the spatial location model.}
\item{MatScaleSp}{The design matrix for the spatial scale model.}
}
}
\description{
This is a helper function that prepares data of sliding block maxima as well
as covariates and response surfaces that are considered such that optimisation
gets more efficient.
}
\examples{
##### generate some data #####
blcksz <- 90
xx <- sapply(c(0.1, 0.4), function(x) evd::rgpd(10*blcksz, shape = x))

yy <- data.frame(xx) \%>\% tidyr::pivot_longer( 1:2, names_to = "Station", values_to = "Obs")

# define a temporal covariate that is constant over a block of length blcksz
temp_cvrt <- rep(1:10/10, each = blcksz)[1:(9*blcksz + 1)]

bms <- get_uniq_bm(yy, blcksz, temp_cvrt = temp_cvrt)
###############################

prep4spatml(loc.sp.form = ~ lon, scale.sp.form = ~ lon +lat,
loc.temp.form = ~ GMST, data = bms, spat.cov = data.frame( lon = c(1,2), lat = c(2,4)))

}
