% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFsl-package.R
\docType{package}
\name{slbm-package}
\alias{slbm}
\alias{slbm-package}
\title{slbm: Extreme Value Statistic based on Sliding Block Maxima}
\description{
This package allows to fit Extreme Value Distributions based on sliding block maxima. Several Models can be fitted, including univariate time series (possibly non-stationary), Intensity Duration Frequency models and Spatial models.
}
\keyword{internal}
