% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mda_data.R
\name{theo_pars}
\alias{theo_pars}
\title{Theoretical coefficients of linear GEV parameter model}
\usage{
theo_pars(
  blcksz,
  loc0 = 0,
  scale0 = 1,
  loclat = 0,
  scalelat = 0,
  loclon = 0,
  scalelon = 0,
  locele = 0,
  scaleele = 0,
  loctemp = 0,
  scaletemp = 0,
  shape = 0.2
)
}
\arguments{
\item{blcksz}{The blocksize $r$ for which block maxima are computed.}

\item{loc0}{The constant coefficient in \eqn{\mu(t,d)}.}

\item{scale0}{The constant coefficient in \eqn{\sigma(t,d)}.}

\item{loclat}{The coefficient in \eqn{\mu(t,d)} belonging to latitude.}

\item{scalelat}{The coefficient in \eqn{\sigma(t,d)} belonging to latitude.}

\item{loclon}{The coefficient in \eqn{\mu(t,d)} belonging to longitude.}

\item{scalelon}{The coefficient in \eqn{\sigma(t,d)} belonging to longitude.}

\item{locele}{The coefficient in \eqn{\mu(t,d)} belonging to elevation.}

\item{scaleele}{The coefficient in \eqn{\sigma(t,d)} belonging to elevaiton.}

\item{loctemp}{The coefficient in \eqn{\mu(t,d)} belonging to the temporal covariate.}

\item{scaletemp}{The coefficient in \eqn{\sigma(t,d)} belonging to the temporal covariate.}

\item{shape}{The shape parameter.}
}
\value{
A vector with approximate coefficient values (only those which are not 0).
}
\description{
Computes approximate theoretical values of the coefficients of the linear models
of GEV parameters for the case where block maxima are of the form
\eqn{\max{\sigma(t,d)X_1 + \mu(t,d), ..., \sigma(t,d)X_r + \mu(t,d) }},
with \eqn{\sigma(t,d)} and \eqn{\mu(t,d)} linear functions of spatial and
temporal covariates, and \eqn{X_i} GPD distributed random variables.
}
\examples{
theo_pars(90, loc0 = 10, scale0 = 1, scalelon = 0.2,  shape = -0.2)
}
