% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conc_bm_id.R
\name{compute_conc_bm_id}
\alias{compute_conc_bm_id}
\title{#' Compute concatenated block maxima}
\usage{
compute_conc_bm_id(agg_df, nlo = 3, resolution = "hourly", seasonlength = 365)
}
\arguments{
\item{agg_df}{Dataframe with values of intensity duration process; output of
function 'fun_aggregate2df'}

\item{nlo}{number of Indices (years/seasons ...) that make up the test set}

\item{resolution}{Resolution of the data}

\item{seasonlength}{The length of the season of the input data. E.g., if observations
are from the whole year, it is 365; if observations are from JJA only, it is 92.}
}
\value{
A tibble with
\itemize{
\item conc_slbm: sliding block maxima as obtained when observations with the i
}
}
\description{
Compute sliding block maxima of the intensity duration process
for all possible compositions of years
arising from cross validation with test set of size given in 'nlo'
}
\examples{
dates <- seq(as.POSIXct("2000-01-01 00:00:00"),
as.POSIXct("2005-12-31 23:00:00"),by = 'hour')
prec <- rgamma(length(dates), shape = 0.1)
ExampleData <- data.frame(datetime = dates, prec = prec)

aggdat <-  fun_aggregate2df(ExampleData, ds = c(1,5,8))
compute_conc_bm_id(aggdat, nlo = 2)

}
