% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fun_wtrend.R
\name{est_var_univ}
\alias{est_var_univ}
\title{GEV parameter covariance matrix for sliding blocks ML estimator}
\usage{
est_var_univ(
  orig_slbm,
  est_par,
  blcksz,
  temp.cov = NULL,
  type = "shift",
  varmeth = "both"
)
}
\arguments{
\item{orig_slbm}{Full sample of sliding block maxima}

\item{est_par}{Output of \code{\link{fit_gev_univ}}.}

\item{blcksz}{The blocksize used.}

\item{temp.cov}{The temporal covariate (must be same length as \code{orig_slbm}).}

\item{type}{Either 'shift', 'scale' or 'stationary', depending if one assumes
that obervations are
shifted or scaled w.r.t. the covariate or stationary (for more details, see
\code{\link[slbm]{fit_gev_univ}}).}

\item{varmeth}{One of 'V', 'V2' or 'both'. Determines the method used for estimating
the covariance matrix.}
}
\value{

}
\description{
GEV parameter covariance matrix for sliding blocks ML estimator
}
\examples{
### simulate some data with a linear trend
xx <- evd::rgpd(90*100, shape = 0.2) + 2*rep(1:100/100, each = 90)
## temporal covariate for sliding BM
temp_cvrt <- rep(1:100/100, each = 90)[1:8911]

## compute sample of uniuqe sliding BM
bms <- get_uniq_bm(xx, 90, temp_cvrt = temp_cvrt, looplastblock = FALSE)

## full sample of sliding for estimaing the covariance matrix
slbm <- blockmax(xx, r = 90, "sliding")
estim <- fit_gev_univ(data = bms, type = "shift", hessian = TRUE)
est_var_univ(slbm, est_par = estim, blcksz = 90, temp.cov = temp_cvrt,
varmeth = "both", type = "shift")
}
