% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aggregate2df.R
\name{fun_aggregate2df}
\alias{fun_aggregate2df}
\title{Compute intensity duration process and block maxima for several durations $d$}
\usage{
fun_aggregate2df(data, ds, resolution = "hourly", seasonlength = 365)
}
\arguments{
\item{data}{a data frame or tibble of precipitation data, with a column named
\itemize{
\item datetime: containing the date-time of the observation as POSIXct, POSIXlt object
\item prec: containing the observed precipitation amount
}}

\item{ds}{the durations for which you want to compute the intensity process.
For each $d$, the intensity process is computed, which is the moving average over d observations
of precipitation, computed within distinct years.}

\item{resolution}{The resolution of the data, must be one of "hourly" or "daily"}

\item{seasonlength}{The length of the season of the input data. E.g., if observations
are from the whole year, it is 365; if observations are from JJA only, it is 92.}
}
\value{
returns a tibble with colums
\itemize{
\item aggs: list of tibbles of aggregated observations, Year they belong to and the duration d.
One tibble for each duration $d$
\item djbm: list of tibbles of annual maxima of the intensity duration process, along with
corresponding year and duration. One tibble for each duration $d$.
}
}
\description{
Compute intensity duration process and block maxima for several durations $d$
}
\examples{
dates <- seq(as.POSIXct("2000-01-01 00:00:00"),
as.POSIXct("2001-12-31 23:00:00"),by = 'hour')
prec <- rgamma(length(dates), shape = 0.1)
ExampleData <- data.frame(datetime = dates, prec = prec)

fun_aggregate2df(ExampleData, ds = c(1,2,4,8,16, 24, 48))
}
