% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_d_fit.R
\name{gev.d.fit.sl}
\alias{gev.d.fit.sl}
\title{Fit d-GEV based on sliding BM}
\usage{
gev.d.fit.sl(
  agg_bm,
  mult_sc = FALSE,
  dur_offset = FALSE,
  int_offset = FALSE,
  sigma0link = make.link("log"),
  optimMethod = "Nelder-Mead",
  Maxit = 1000,
  ...
)
}
\arguments{
\item{agg_bm}{Output of get_agg_bm, i.e. a tibble containing the sample of disjoint
and sliding block maxima of the ID process,
for different durations $d$}

\item{mult_sc}{logical. Whether to fit a multi-scale parameter eta2}

\item{dur_offset}{logical. Whether to fit a duration offset parameter theta}

\item{int_offset}{logical. Whether to fit an intensity offset parameter tau}

\item{sigma0link}{The link function for \eqn{\sigma_0}. Default is \code{ make.link("log")}.}

\item{optimMethod}{The optimisation method used in \code{\link[=optim]{optim()}}.}

\item{Maxit}{Passed to the control argument of \code{\link[=optim]{optim()}}:
maximum number of iterations performed during the optimisation.}

\item{...}{Further arguments that can be passed to the control argument of \code{\link[=optim]{optim()}}.}
}
\value{
Returns a list with the following arguments:
\itemize{
\item mle : A dataframe containing the estimated parameters based on disjoint and sliding block
maxima as specified in column \code{estimator}
\item conv: logical; whether optimisation succeeded
\item nllh: value of the negative log-Likelihoods (disjoint and sliding)
\item counts: Number of calls to the negative log-Likelihood
\item dur_offset: logical; whether the duration offset was fitted
\item mult_sc: logical; whether the duration offset was fitted
}
}
\description{
Fit d-GEV based on sliding BM
}
\examples{
dates <- seq(as.POSIXct("2000-01-01 00:00:00"),
as.POSIXct("2020-12-31 23:00:00"),by = 'hour')
prec <- rgamma(length(dates), shape = 0.1)
example_data <- data.frame(datetime = dates, prec = prec)

agbm <- get_agg_bm(example_data, ds = c(1,2,4,8,16, 24, 48))
gev.d.fit.sl( agbm)
}
