% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_boot_sl.R
\name{sample_boot_sl}
\alias{sample_boot_sl}
\title{Studentized bootstrap repetition}
\usage{
sample_boot_sl(
  nKblocks,
  sluniq,
  slorig,
  blcksz,
  start_vals = NULL,
  type = "stationary",
  scale.link = make.link("identity"),
  seed,
  reltol = 1e-08,
  estimate_RL = TRUE,
  Tyrl = c(50, 100),
  ref_gmst = NULL,
  varmeth = "V2",
  chain = TRUE,
  shape_constraint = -0.9,
  ...
)
}
\arguments{
\item{nKblocks}{Indices of blocks that make up the bootstrap sample}

\item{sluniq}{Tibble with column \code{Station} and \code{uniq_data}, where the latter
contains a tibble with columns \code{Kblockind}, \code{slbm}, \code{n} giving the index of the
K-block, the unqiue sliding block maxima within that block and their frequency, respectively.}

\item{slorig}{The complete sliding block maxima sample for the looped blocks.}

\item{blcksz}{The blocksize parameter.}

\item{start_vals}{Starting values.}

\item{type}{The type of the model to fit: one of \code{stationary}, \code{shift} or \code{scale}}

\item{scale.link}{The link function for the scale parameter.}

\item{seed}{A seed for the sampling of blocks.}

\item{reltol}{Passed to optim.}

\item{estimate_RL}{Must be TRUE. GEV parameter Estimation not implemented atm.}

\item{Tyrl}{The return periods for which RLs are estimated}

\item{ref_gmst}{Vector of reference values of the temporal covariate for which
RLs are estimated. Ignored when type is 'stationary'.}

\item{varmeth}{The method used to estimate the covariance matrix of GEV parameters: 'V' or 'V2'.}

\item{chain}{logical and only relevant for non-stationary models: Whether the method for
covariance estimation should be based on chain rule}

\item{...}{Further arguments passed to \code{fit_gev_univ}. When \code{type = scale}, one must
pass the logical argument \code{rel_trend} which specifies the parametrisation of the scale model.}
}
\value{
A list containing
\itemize{
\item \code{est_boot} the parameter estimates of the bootstrap sample
and one or both of the following:
\item \code{estvarV} the estimated covariance matrix of the parameter estimates (method 1)
\item \code{estvarV2} the estimated covariance matrix of the parameter estimates (method 1)
}

when \code{stimate_RL = FALSE}. If \code{estimate_RL = TRUE}, a list with the following
components is returned:
\itemize{
\item \code{rlboot} the estimated RLs
and one or both of the following:
\item \code{rlvarV} the estimated variances of the RLs (method 1)
\item \code{rlvarV2} the estimated variances of the RLs (method 2)
}
}
\description{
Studentized bootstrap repetition
}
\examples{
# none
}
